refactor: routieSpace 관련 api type 파일 분리
