+++ frontend/src/domains/routieSpace/types/api.types.ts
@@ -1,0 +1,22 @@
+interface CreateRoutieResponseType {
+  routieSpaceIdentifier: string;
+}
+
+interface GetRoutieSpaceResponseType {
+  name: string;
+}
+
+interface EditRoutieSpaceNameRequestType {
+  name: string;
+}
+
+interface EditRoutieSpaceNameResponseType {
+  name: string;
+}
+
+export type {
+  CreateRoutieResponseType,
+  GetRoutieSpaceResponseType,
+  EditRoutieSpaceNameRequestType,
+  EditRoutieSpaceNameResponseType,
+};
