+++ frontend/src/domains/routieSpace/types/routieSpace.types.ts
@@ -1,17 +1,5 @@
 import { ERROR_MESSAGE } from '../constants/routieSpace';
 
-interface CreateRoutieResponseType {
-  routieSpaceIdentifier: string;
-}
-
-interface GetRoutieSpaceResponseType {
-  name: string;
-}
-
-interface EditRoutieSpaceNameResponseType {
-  name: string;
-}
-
 type ERROR_CASE = keyof typeof ERROR_MESSAGE;
 
 interface UseRoutieSpaceNameReturn {
@@ -25,10 +13,4 @@ interface UseRoutieSpaceNameReturn {
   handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
-export type {
-  CreateRoutieResponseType,
-  GetRoutieSpaceResponseType,
-  EditRoutieSpaceNameResponseType,
-  ERROR_CASE,
-  UseRoutieSpaceNameReturn,
-};
+export type { ERROR_CASE, UseRoutieSpaceNameReturn };
